Pass all pikaday options in separate prop
Adds a `pikadayOptions` prop that is an object that is passed straight
to the Pikaday plugin. This allows users to configure pikaday however
they need to, while remaining flexible enough to not need updates when
pikaday changes.

Also adds `babel-preset-stage-2` to splat the pikadayOptions prop
without needing to do a merge.